P6.07-Model.java
Removed methods
